Reproduce test case where we see an AIOOB with a DTR but not with a AlwaysRefreshDTR,
Also separate it out into a different class directly into the Facet package

Change constructor to be protected instead of public to allow subclasses to be able to modify it.
